Combined the two testing files for mixins.js
