--- testing/testing 2.txt
@@ -1,17 +1,0 @@
-<<<<<<< HEAD
-Methods to test:
-
-mixins.js
-
-getWeekString
-getWeekNumber
-weeksFromToday
-scrollScheduleIntoView
-
-dateToString
-parseEndTime
-getTimeString
-formatTime
-=======
-testing
->>>>>>> eebb3ac5d42f5e084065304ad9acaaa437fc247e
