--- testing/testing.txt
@@ -1,13 +1,0 @@
-Methods to test:
-
-mixins.js
-
-getWeekString
-getWeekNumber
-weeksFromToday
-scrollScheduleIntoView
-
-dateToString
-parseEndTime
-getTimeString
-formatTime
